+++ src/popup/App.js
@@ -3,8 +3,15 @@ import { getAllRouter as AllRouter } from './router';
 import IdleTimer from 'react-idle-timer'
 import { sendMsg } from '../utils/commonMsg';
 import { WALLET_RESET_LAST_ACTIVE_TIME } from '../constant/types';
-import { useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import cls from "classnames"
+import { getFeeRecom, getScamList } from '../background/api';
+import { updateRecomFee } from '../reducers/cache';
+import { useDispatch, useSelector } from 'react-redux';
+import { getLocal } from '../background/localStorage';
+import { RECOMMOND_FEE, SCAM_LIST } from '../constant/storageKey';
+import { updateScamList } from '../reducers/accountReducer';
+import { NET_CONFIG_TYPE } from '../constant/walletType';
 
 function setLastActiveTime(){ 
   sendMsg({
@@ -13,6 +20,9 @@ function setLastActiveTime(){
 }
 
 function App() {
+  const dispatch = useDispatch()
+  const netConfig = useSelector(state => state.network.currentConfig)
+
   const [showFullStatus,setShowFullStatus] = useState(false)
   useEffect(()=>{
     const url = new URL(window.location.href); 
@@ -25,6 +35,57 @@ function App() {
       setShowFullStatus(true)
     }
   },[window.location.href])
+
+  const getLocalFeeList = useCallback(() => {
+    let localFeeList = getLocal(RECOMMOND_FEE)
+    if (localFeeList) {
+      let feeList = JSON.parse(localFeeList)
+      dispatch(updateRecomFee(feeList))
+    }
+  }, [])
+
+  const fetchFeeData = useCallback(async () => {
+    let feeRecom = await getFeeRecom()
+    if (feeRecom.length > 0) {
+      dispatch(updateRecomFee(feeRecom))
+    }else{
+      getLocalFeeList()
+    }
+  }, [])
+  const getLocalScamList = useCallback(() => {
+    let localScamList = getLocal(SCAM_LIST)
+    if (localScamList) {
+      let scamList = JSON.parse(localScamList)
+      dispatch(updateScamList(scamList))
+    }
+  }, [])
+  const fetchScamList = useCallback(async()=>{
+    let scamList = await getScamList()
+    if (scamList.length > 0) {
+      dispatch(updateScamList(scamList))
+    }
+  },[])
+
+  const initNetData = useCallback(()=>{
+    fetchFeeData()
+    const netType = netConfig?.netType
+    if(netType === NET_CONFIG_TYPE.Mainnet){
+        fetchScamList()  
+    }
+  },[netConfig])
+
+  useEffect(()=>{
+    const netType = netConfig?.netType
+    if(netType !== NET_CONFIG_TYPE.Mainnet){
+        dispatch(updateScamList([]))
+    }else{
+      getLocalScamList()
+    }
+  },[netConfig])
+
+  useEffect(() => {
+    initNetData()
+  }, [])
   
   return (
     <div className="App">